fix clear the text input error and reverse all changes related to implement updatepost;
